New Features [Driver]: - Adding a new "endpoints" architecture for the driver that allows multiple plugin DLL's (including some in other processes accessed via IPC) to receive the exact same command-stream as is being sent to the physical GPU device. This allows for robust recording/playback, profiling, and differential testing across many different API's, hardware/software rasterizers, and devices compared to before. - Adding support for x64 targets for the main driver (previously only x86 was ever tested), and for most of its DLL's and some of its companion programs. - Uploading many new TestProjects (into the new Software_d3d9_Driver/TestProjects directory). These small test programs are used for exercising individual bits of renderer functionality one at a time and are useful for testing purposes to make sure that various features work as expected. More will be added in the future as well for testing other features. - Added new support to the driver to be able to handle rendering D3DFVF_XYZRHW / D3DDECLUSAGE_POSITIONT vertices (ie. those with pretransformed vertices).
Updated Features [Driver]:
- Reworked some of the vertex-buffer and index-buffer uploading code to now perform lazy uploads only when the vertex/index buffer is actually used at draw-call-time, rather than uploading vertex and index buffers that are never being drawn.

Bugfixes [Driver]:
- Fixed a bug in the shader compiler that was producing incorrect source modifiers for frc and replicate-swizzle (exp, expp, log, logp, rcp, rsq) operations.
- Fixed a bug in the shader compiler that was using a wrong register swizzle when decomposing the mad instruction.
- Fixed a bug where render targets with dimensions that are not the screen dimension (ie. 256x256 or 512x512 render targets) would not be cleared properly, resulting in some memory corruption.
- Fixed a bug where render target textures would be allocated with the incorrect GPU memory alignment (they would previously get the texture alignment rather than the rendertarget alignment).
- Fixed a bug when using DrawPrimitive() or DrawIndexedPrimitive() with a starting offset of 8192 or more.
- Added a driver workaround for an app bug in which an app was setting SetRenderState(D3DRS_ALPHAFUNC, 0) (and in this case 0 does not map to a valid D3DCMP enum).
- Fixed a GPU memory leak when a surface was released.

License info (TestProjects/SimpleDXTTest1/TestDXT5.dds):
- This DXT5 image is under the CC BY-SA 3.0 license. I have modified the image by cropping and resizing it from the original 800x600 to the new 256x256 resolution, and also converting the file from PNG to DDS format.
- Original source image and its original license is from Wikipedia here: https://commons.wikimedia.org/wiki/File:PNG_transparency_demonstration_1.png

License info (other test projects assets):
- The other TestProjects image assets are all either public-domain images, or are my own creations (and they are also public-domain and others may use them for whatever they wish).

+++ CPUDrivenGPUTest2/Software_d3d9_Driver/Software_d3d9_Driver/ShaderAnalysis.cpp
@@ -316,7 +316,7 @@ static inline void ResolveDstParameter(const DWORD*& bytecode, ShaderInfo& shade
 
 static inline void AddIfNotPresent(std::vector<unsigned short>& vals, const unsigned short val)
 {
-	const unsigned numVals = vals.size();
+	const unsigned numVals = (const unsigned)vals.size();
 	for (unsigned x = 0; x < numVals; ++x)
 	{
 		if (vals[x] == val)
@@ -968,7 +968,7 @@ static inline const bool ParseCustomOpcode(const D3DSHADER_INSTRUCTION_OPCODE_TY
 		break;
 	case D3DSIO_END:
 		if (print) dprintf(shaderInfo, "\n");
-		shaderInfo.shaderLengthDWORDs = shaderMemory - shaderInfo.initialBytecodeToken;
+		shaderInfo.shaderLengthDWORDs = (const unsigned)(shaderMemory - shaderInfo.initialBytecodeToken);
 		return true;
 	default:
 #ifdef _DEBUG
@@ -1327,7 +1327,7 @@ void outputRegisterWriteTracker::SetAddOrSkip(const D3DSHADER_PARAM_REGISTER_TYP
 
 void outputRegisterWriteTracker::AddOutputRegisterListingToShaderInfo(ShaderInfo& outShaderInfo)
 {
-	const unsigned numWrittenOutputRegisters = registerIdentifiers.size();
+	const unsigned numWrittenOutputRegisters = (const unsigned)registerIdentifiers.size();
 	for (unsigned x = 0; x < numWrittenOutputRegisters; ++x)
 	{
 		const outputRegisterIdentifier& registerIdentifier = registerIdentifiers[x];
@@ -1408,9 +1408,9 @@ void __stdcall PrintShaderStatsToString(char (&outBuffer)[1024], const ShaderInf
 		shaderInfoIn.numArithInstructions, shaderInfoIn.numTexInstructions,
 		shaderInfoIn.numUniqueTempRegistersUsed, MAX_NUM_TEMP_REGISTERS,
 		shaderInfoIn.numUniqueInputRegistersUsed, shaderInfoIn.isPixelShader ? ( (shaderInfoIn.shaderMajorVersion < 2) ? 8 : 10) : 16,
-		shaderInfoIn.usedConstantsF.size(),
-		shaderInfoIn.usedConstantsI.size(), MAX_NUM_CONST_INT_REGISTERS,
-		shaderInfoIn.usedConstantsB.size(), MAX_NUM_CONST_BOOL_REGISTERS,
+		(const unsigned)shaderInfoIn.usedConstantsF.size(),
+		(const unsigned)shaderInfoIn.usedConstantsI.size(), MAX_NUM_CONST_INT_REGISTERS,
+		(const unsigned)shaderInfoIn.usedConstantsB.size(), MAX_NUM_CONST_BOOL_REGISTERS,
 		shaderInfoIn.usesDynamicBranching ? "TRUE" : "FALSE",
 		shaderInfoIn.usesFunctionCalls ? "TRUE" : "FALSE",
 		shaderInfoIn.usesIntToFloatConversions ? "TRUE" : "FALSE",
